+++ ESI.NET/Enumerations/RoutePreference.cs
@@ -5,7 +5,7 @@ namespace ESI.NET.Enumerations
 {
     /// <summary>Routing preference for <see cref="ESI.NET.Logic.RoutesLogic.Map"/>.</summary>
     [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
-    public enum RoutesFlag
+    public enum RoutePreference
     {
         [EnumMember(Value = "Shorter")]    /**/ Shorter,
         [EnumMember(Value = "Safer")]      /**/ Safer,